# Pool buffers when decoding