--- a/rpc/transport/pipe.go
+++ b/rpc/transport/pipe.go
@@ -72,6 +72,8 @@ func (p *pipe) Decode() (*capnp.Message, error) {
 
 }
 
+func (*pipe) ReleaseMessage(*capnp.Message) {}
+
 func (p *pipe) Close() error {
 	close(p.closed)
 	syncutil.With(&p.sendMu, func() {