@zenhack 2022-12-01
It's spending a bunch of time in Pool.Put (close to the increase); I assume it's the bit where it zeros the buffer. Perhaps the zeroing of memory from memory from make() is faster, and that's the difference?

I'd like to hear what this does to @lthibault's real workload before fussing with it too much.