@lthibault 2022-12-01
>Perhaps the zeroing of memory from memory from make() is faster, and that's the difference?

I don't have a reference on hand, but I looked into this recently and can confirm this to be the case.

Perhaps we shouldn't zero memory for now, and instead make a note of the security implications in a docstring?