@zenhack 2022-12-01
> Perhaps we shouldn't zero memory for now, and instead make a note of the security implications in a docstring?

I'm *really* against this approach. I'd be ok with a flag to explicitly turn *off* zeroing, but we should try to be secure by default. Most of the stuff I'm using this for I wouldn't be willing to make that trade-off, and wouldn't want it made that way behind my back.

But before we added such a flag, I'd want to validate that this is actually an issue in prod; The things I can think of that would make `make()` faster than `memset()` (which the compiler should be smart enough to optimize that loop into) are things like mapping in the zero page for new memory, which avoids needing to actually touch the memory at all -- but that won't hold up in a steady-state scenario. Basically, I don't want to give up security/robustness when there are good reasons to believe this could just be the synthetic benchmark not measuring exactly what we're interested in.