@lthibault 2022-12-01
> > Perhaps we shouldn't zero memory for now, and instead make a note of the security implications in a docstring?
> 
> I'm _really_ against this approach. I'd be ok with a flag to explicitly turn _off_ zeroing, but we should try to be secure by default. Most of the stuff I'm using this for I wouldn't be willing to make that trade-off, and wouldn't want it made that way behind my back.

That settles it, then. 👍  Let's grab some good data and take it from there.